corrects style
