+++ src/main/webapp/index.html
@@ -9,7 +9,7 @@
   </head>
   <body onload="makeGraph()">
       <p> Hello </p>
-    <div id="chart_div" style="width: 900px; height: 500px;"></div>
-    <div id="chart_div2" style="width: 900px; height: 500px;"></div>
+    <div id="hours-chart" style="width: 900px; height: 500px;"></div>
+    <div id="difficulty-chart" style="width: 900px; height: 500px;"></div>
   </body>
 </html>
